Small fix for type

+++ app/helpers/application_helper.rb
@@ -11,7 +11,7 @@ module ApplicationHelper
   include ActionView::Helpers::DateHelper
   include Kernel
 
-  sig { params(secs: Integer).returns(String) }
+  sig { params(secs: Float).returns(String) }
   def duration_of_time_in_words(secs)
     distance_of_time_in_words(0, secs, include_seconds: true)
   end
